% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{setup_github_personal_access_token}
\alias{setup_github_personal_access_token}
\title{Setup Github personal access token after setting up username and email}
\usage{
setup_github_personal_access_token()
}
\value{
None
}
\description{
Setup Github personal access token after setting up username and email
}
