% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{setup_github_username_email}
\alias{setup_github_username_email}
\title{Setup Github username and email}
\usage{
setup_github_username_email()
}
\value{
None
}
\description{
Setup Github username and email
}
