% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{update_github_credentials}
\alias{update_github_credentials}
\title{Update Github credentials
Run this after setting up username, email and personal access token. Or
when github push stops working in RStudio.}
\usage{
update_github_credentials()
}
\value{
None
}
\description{
Update Github credentials
Run this after setting up username, email and personal access token. Or
when github push stops working in RStudio.
}
